Fix -Wchar-subscripts and two -Wsign-compare warnings
sliba.c:1261: base64_encode_table elements are char (signed on this
platform). Indexing into base64_decode_table[256] with a signed char
risks a negative index for values >= 128. Cast to (unsigned char) before
indexing — base64 characters are all ASCII so this is a no-op in
practice, but it is the correct type.

slib.c:2341: j is a non-negative buffer index; cast (j+1) to size_t so
both sides of the comparison have the same unsigned type.

slibu.c:1447: n is long intentionally — the negative case is checked
two lines above with (n < 0). Cast sizeof(buffer) to long instead; the
buffer is 2048 bytes, well within long range on all targets.

+++ slib.c
@@ -2338,7 +2338,7 @@ LISP vload(char *ofname,long cflag,long rflag)
  c = getc(f);
  while((c == '#') || (c == ';'))
    {while(((c = getc(f)) != EOF) && (c != '\n'))
-      if ((j+1)<sizeof(buffer))
+      if ((size_t)(j+1)<sizeof(buffer))
 	{buffer[j] = c;
 	 buffer[++j] = 0;}
     if (c == '\n')
